+++ python_modules/libraries/dagster-fivetran/dagster_fivetran_tests/beta/test_translator.py
@@ -10,7 +10,7 @@ from dagster_fivetran import (
     FivetranWorkspace,
 )
 
-from dagster_fivetran_tests.experimental.conftest import (
+from dagster_fivetran_tests.beta.conftest import (
     TEST_ACCOUNT_ID,
     TEST_API_KEY,
     TEST_API_SECRET,
